chore(perf): allow sharding efps tests
