+++ perf/efps/.eslintrc.cjs
@@ -1,0 +1,9 @@
+const path = require('node:path')
+
+const ROOT_PATH = path.resolve(__dirname, '../../..')
+
+module.exports = {
+  rules: {
+    'import/no-extraneous-dependencies': ['error', {packageDir: [ROOT_PATH, __dirname]}],
+  },
+}
