+++ perf/efps/package.json
@@ -12,7 +12,8 @@
   },
   "scripts": {
     "start": "node --import @swc-node/register/esm-register ./index.ts",
-    "test": "npm start"
+    "test": "npm start",
+    "write:report": "node --import @swc-node/register/esm-register ./formatefpsResult.ts"
   },
   "devDependencies": {
     "@sanity/client": "^6.22.1",
@@ -20,6 +21,7 @@
     "@types/react": "^18.3.3",
     "@types/react-dom": "^18.3.0",
     "@types/serve-handler": "^6.1.4",
+    "@types/yargs": "^17.0.7",
     "@vitejs/plugin-react": "^4.3.1",
     "chalk": "^4.1.2",
     "cli-table3": "^0.6.5",
@@ -33,6 +35,7 @@
     "sanity": "workspace:*",
     "serve-handler": "^6.1.5",
     "source-map": "^0.7.4",
-    "vite": "^5.4.2"
+    "vite": "^5.4.2",
+    "yargs": "17.3.0"
   }
 }
